+++ pennylane/templates/subroutines/single_excitation_unitary.py
@@ -12,37 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 r"""
-Contains the ``SingleExcitationUnitary`` template.
+Contains the SingleExcitationUnitary template.
 """
+# pylint: disable-msg=too-many-branches,too-many-arguments,protected-access
 import pennylane as qml
 from pennylane import numpy as np
-
-# pylint: disable-msg=too-many-branches,too-many-arguments,protected-access
-from pennylane.ops import CNOT, RX, RZ, Hadamard
-from pennylane.templates.decorator import template
-from pennylane.wires import Wires
+from pennylane.operation import Operation, AnyWires
 
 
-def _preprocess(weight, wires):
-    """Validate and pre-process inputs as follows:
-
-    * Check the shape of the weights tensor.
-    * Check that there are at least 2 wires.
-
-    Args:
-        weight (tensor_like): trainable parameters of the template
-        wires (Wires): wires that template acts on
-    """
-    if len(wires) < 2:
-        raise ValueError("expected at least two wires; got {}".format(len(wires)))
-
-    shape = qml.math.shape(weight)
-    if shape != ():
-        raise ValueError(f"Weight must be a scalar tensor {()}; got shape {shape}.")
-
-
-@template
-def SingleExcitationUnitary(weight, wires=None):
+class SingleExcitationUnitary(Operation):
     r"""Circuit to exponentiate the tensor product of Pauli matrices representing the
     single-excitation operator entering the Unitary Coupled-Cluster Singles
     and Doubles (UCCSD) ansatz. UCCSD is a VQE ansatz commonly used to run quantum
@@ -92,14 +70,11 @@ def SingleExcitationUnitary(weight, wires=None):
 
     Args:
         weight (float): angle :math:`\theta` entering the Z rotation acting on wire ``p``
-        wires (Iterable or Wires): Wires that the template acts on.
+        wires (Iterable): Wires that the template acts on.
             The wires represent the subset of orbitals in the interval ``[r, p]``. Must be of
             minimum length 2. The first wire is interpreted as ``r`` and the last wire as ``p``.
             Wires in between are acted on with CNOT gates to compute the parity of the set
             of qubits.
-
-    Raises:
-        ValueError: if inputs do not have the correct format
 
     .. UsageDetails::
 
@@ -136,56 +111,75 @@ def SingleExcitationUnitary(weight, wires=None):
 
     """
 
-    wires = Wires(wires)
-    _preprocess(weight, wires)
+    num_params = 1
+    num_wires = AnyWires
+    par_domain = "A"
 
-    # Interpret first and last wire as r and p
-    r = wires[0]
-    p = wires[-1]
+    def __init__(self, weight, wires=None, do_queue=True):
+        if len(wires) < 2:
+            raise ValueError("expected at least two wires; got {}".format(len(wires)))
 
-    # Sequence of the wires entering the CNOTs between wires 'r' and 'p'
-    set_cnot_wires = [wires.subset([l, l + 1]) for l in range(len(wires) - 1)]
+        shape = qml.math.shape(weight)
+        if shape != ():
+            raise ValueError(f"Weight must be a scalar tensor {()}; got shape {shape}.")
 
-    # ------------------------------------------------------------------
-    # Apply the first layer
+        super().__init__(weight, wires=wires, do_queue=do_queue)
 
-    # U_1, U_2 acting on wires 'r' and 'p'
-    RX(-np.pi / 2, wires=r)
-    Hadamard(wires=p)
+    def expand(self):
 
-    # Applying CNOTs between wires 'r' and 'p'
-    for cnot_wires in set_cnot_wires:
-        CNOT(wires=cnot_wires)
+        weight = self.parameters[0]
 
-    # Z rotation acting on wire 'p'
-    RZ(weight / 2, wires=p)
+        # Interpret first and last wire as r and p
+        r = self.wires[0]
+        p = self.wires[-1]
 
-    # Applying CNOTs in reverse order
-    for cnot_wires in reversed(set_cnot_wires):
-        CNOT(wires=cnot_wires)
+        # Sequence of the wires entering the CNOTs between wires 'r' and 'p'
+        set_cnot_wires = [self.wires[l: l + 2] for l in range(len(self.wires) - 1)]
 
-    # U_1^+, U_2^+ acting on wires 'r' and 'p'
-    RX(np.pi / 2, wires=r)
-    Hadamard(wires=p)
+        with qml.tape.QuantumTape() as tape:
 
-    # ------------------------------------------------------------------
-    # Apply the second layer
+            # ------------------------------------------------------------------
+            # Apply the first layer
 
-    # U_1, U_2 acting on wires 'r' and 'p'
-    Hadamard(wires=r)
-    RX(-np.pi / 2, wires=p)
+            # U_1, U_2 acting on wires 'r' and 'p'
+            qml.RX(-np.pi / 2, wires=r)
+            qml.Hadamard(wires=p)
 
-    # Applying CNOTs between wires 'r' and 'p'
-    for cnot_wires in set_cnot_wires:
-        CNOT(wires=cnot_wires)
+            # Applying CNOTs between wires 'r' and 'p'
+            for cnot_wires in set_cnot_wires:
+                qml.CNOT(wires=cnot_wires)
 
-    # Z rotation acting on wire 'p'
-    RZ(-weight / 2, wires=p)
+            # Z rotation acting on wire 'p'
+            qml.RZ(weight / 2, wires=p)
 
-    # Applying CNOTs in reverse order
-    for cnot_wires in reversed(set_cnot_wires):
-        CNOT(wires=cnot_wires)
+            # Applying CNOTs in reverse order
+            for cnot_wires in reversed(set_cnot_wires):
+                qml.CNOT(wires=cnot_wires)
 
-    # U_1^+, U_2^+ acting on wires 'r' and 'p'
-    Hadamard(wires=r)
-    RX(np.pi / 2, wires=p)
+            # U_1^+, U_2^+ acting on wires 'r' and 'p'
+            qml.RX(np.pi / 2, wires=r)
+            qml.Hadamard(wires=p)
+
+            # ------------------------------------------------------------------
+            # Apply the second layer
+
+            # U_1, U_2 acting on wires 'r' and 'p'
+            qml.Hadamard(wires=r)
+            qml.RX(-np.pi / 2, wires=p)
+
+            # Applying CNOTs between wires 'r' and 'p'
+            for cnot_wires in set_cnot_wires:
+                qml.CNOT(wires=cnot_wires)
+
+            # Z rotation acting on wire 'p'
+            qml.RZ(-weight / 2, wires=p)
+
+            # Applying CNOTs in reverse order
+            for cnot_wires in reversed(set_cnot_wires):
+                qml.CNOT(wires=cnot_wires)
+
+            # U_1^+, U_2^+ acting on wires 'r' and 'p'
+            qml.Hadamard(wires=r)
+            qml.RX(np.pi / 2, wires=p)
+
+        return tape
